+++ Dev/CondorcetDocumentationGenerator/CondorcetDocAttributes/Example.php
@@ -15,12 +15,7 @@ use Attribute;
 #[Attribute(Attribute::TARGET_METHOD | Attribute::IS_REPEATABLE)]
 class Example
 {
-    public readonly string $name;
-    public readonly string $link;
-
-    public function __construct(string $name, string $link)
+    public function __construct(public readonly string $name, public readonly string $link)
     {
-        $this->name = $name;
-        $this->link = $link;
     }
 }